requirements.txt
# requirements.txt
python-telegram-bot
aiohttp
python-dotenv